more code organization
