+++ examples/full-page/src/pages/Home/index.tsx
@@ -1,8 +1,7 @@
 import { FullPageView } from "@stripe/ui-extension-sdk/ui";
-import { useReducer } from "react";
 import { Tab, Tabs } from "@stripe/ui-extension-sdk/ui/next";
+import { useRoute } from "@stripe/ui-extension-sdk/navigation";
 
-import { useHomeQuery } from "@/data";
 import { useQueuedToast } from "@/hooks/useQueuedToast";
 import { GrantPointsDrawer } from "./drawers/GrantPointsDrawer";
 import { EditMemberDrawer } from "./drawers/EditMemberDrawer";
@@ -11,90 +10,30 @@ import { ActivityTab } from "./tabs/ActivityTab";
 import { MembersTab } from "./tabs/MembersTab";
 import { OverviewTab } from "./tabs/OverviewTab";
 import { RewardsTab } from "./tabs/RewardsTab";
-import type { Member, Reward } from "@/data";
-import { useRoute } from "@stripe/ui-extension-sdk/navigation";
+import { useHome } from "./useHome";
 
 const DEFAULT_TAB = "overview";
 type HomeTabId = "overview" | "members" | "rewards" | "activity";
-
-type HomeState = {
-  editMemberDrawer: { open: boolean; member: Member | null };
-  grantDrawer: { open: boolean; preselectedMemberId?: string };
-  editRewardDrawer: { open: boolean; reward: Reward | null };
-};
-
-type HomeAction =
-  | { type: "openEditMember"; member: Member }
-  | { type: "closeEditMember" }
-  | { type: "openGrantDrawer"; preselectedMemberId?: string }
-  | { type: "closeGrantDrawer" }
-  | { type: "openEditReward"; reward: Reward }
-  | { type: "closeEditReward" };
-
-const initialHomeState: HomeState = {
-  editMemberDrawer: { open: false, member: null },
-  grantDrawer: { open: false },
-  editRewardDrawer: { open: false, reward: null },
-};
-
-function homeReducer(state: HomeState, action: HomeAction): HomeState {
-  switch (action.type) {
-    case "openEditMember":
-      return {
-        ...state,
-        editMemberDrawer: { open: true, member: action.member },
-      };
-    case "closeEditMember":
-      return {
-        ...state,
-        editMemberDrawer: { open: false, member: null },
-      };
-    case "openGrantDrawer":
-      return {
-        ...state,
-        grantDrawer: {
-          open: true,
-          preselectedMemberId: action.preselectedMemberId,
-        },
-      };
-    case "closeGrantDrawer":
-      return { ...state, grantDrawer: { open: false } };
-    case "openEditReward":
-      return {
-        ...state,
-        editRewardDrawer: { open: true, reward: action.reward },
-      };
-    case "closeEditReward":
-      return {
-        ...state,
-        editRewardDrawer: { open: false, reward: null },
-      };
-  }
-}
 
 export function Home() {
   const {
     route: { routeParams, key },
     setRoute,
   } = useRoute();
-  const { data: homeData } = useHomeQuery();
   const { queueToast } = useQueuedToast();
-  const [state, dispatch] = useReducer(homeReducer, initialHomeState);
+  const {
+    grantDrawer,
+    editMemberDrawer,
+    editRewardDrawer,
+    openEditMember,
+    openGrantDrawer,
+    openEditReward,
+    closeGrantDrawer,
+    closeEditMemberDrawer,
+    closeEditRewardDrawer,
+  } = useHome();
 
   const currentTab = (key === "home" && routeParams.tabId) || DEFAULT_TAB;
-
-  const openEditMember = (memberId: string) => {
-    const member = homeData?.members.find((m) => m.id === memberId);
-    if (member) dispatch({ type: "openEditMember", member });
-  };
-
-  const openGrantDrawer = (preselectedMemberId?: string) =>
-    dispatch({ type: "openGrantDrawer", preselectedMemberId });
-
-  const openEditReward = (rewardId: string) => {
-    const reward = homeData?.rewards.find((r) => r.id === rewardId);
-    if (reward) dispatch({ type: "openEditReward", reward });
-  };
 
   return (
     <FullPageView
@@ -127,13 +66,13 @@ export function Home() {
       </Tabs>
 
       <GrantPointsDrawer
-        shown={state.grantDrawer.open}
+        shown={grantDrawer.open}
         setShown={(shown) => {
-          if (!shown) dispatch({ type: "closeGrantDrawer" });
+          if (!shown) closeGrantDrawer();
         }}
-        preselectedMemberId={state.grantDrawer.preselectedMemberId}
+        preselectedMemberId={grantDrawer.preselectedMemberId}
         onGranted={() => {
-          dispatch({ type: "closeGrantDrawer" });
+          closeGrantDrawer();
           queueToast("Points granted", "success");
         }}
         onGrantError={() => {
@@ -141,18 +80,18 @@ export function Home() {
         }}
       />
       <EditMemberDrawer
-        shown={state.editMemberDrawer.open}
+        shown={editMemberDrawer.open}
         setShown={(shown) => {
-          if (!shown) dispatch({ type: "closeEditMember" });
+          if (!shown) closeEditMemberDrawer();
         }}
-        member={state.editMemberDrawer.member}
+        member={editMemberDrawer.member}
       />
       <EditRewardDrawer
-        shown={state.editRewardDrawer.open}
+        shown={editRewardDrawer.open}
         setShown={(shown) => {
-          if (!shown) dispatch({ type: "closeEditReward" });
+          if (!shown) closeEditRewardDrawer();
         }}
-        reward={state.editRewardDrawer.reward}
+        reward={editRewardDrawer.reward}
       />
     </FullPageView>
   );
