+++ examples/full-page/src/pages/Home/tabs/OverviewTab/index.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import {
   Box,
   Divider,
@@ -8,7 +7,6 @@ import {
   Menu,
   MenuItem,
   OverviewPage,
-  PageModule,
 } from "@stripe/ui-extension-sdk/ui";
 import {
   BarChart,
@@ -17,36 +15,10 @@ import {
 } from "@stripe/ui-extension-sdk/ui/next";
 import { PageModule } from "@stripe/ui-extension-sdk/ui/experimental";
 
-import {
-  pointsToDollars,
-  TierName,
-  useMembersQuery,
-  useOverviewQuery,
-  useSettingsQuery,
-} from "@/data";
 import { formatTimestamp } from "@/utils/date";
 import { formatCurrency, formatPoints } from "@/utils/format";
 import { useQueuedToast } from "@/hooks/useQueuedToast";
-
-// Trend data is chronological Feb 2025 → Feb 2026. Assign years by tracking
-// when months roll back to Jan (the start of 2026 in our data).
-const assignYears = <T extends { week: string }>(data: T[]) => {
-  let year = 2025;
-  let prevMonth = -1;
-  return data.map((d) => {
-    const monthNum = new Date(`${d.week} 2025`).getMonth();
-    if (monthNum < prevMonth) year = 2026;
-    prevMonth = monthNum;
-    return { ...d, date: new Date(`${d.week} ${year}`) };
-  });
-};
-
-const pctChange = (current: number, previous: number): number | null => {
-  if (previous === 0) return current > 0 ? 100 : null;
-  return ((current - previous) / previous) * 100;
-};
-
-const MEMBER_TIERS: TierName[] = ["Bean Counter", "Barista", "Roastmaster"];
+import { useOverviewTab } from "./useOverviewTab";
 
 const GrowthIndicator = ({ pct }: { pct: number | null }) => {
   if (pct === null) return null;
@@ -66,11 +38,20 @@ const GrowthIndicator = ({ pct }: { pct: number | null }) => {
 };
 
 export function OverviewTab() {
-  const { data, isLoading, isError, error } = useOverviewQuery();
-  const { data: members } = useMembersQuery();
-  const { data: settings } = useSettingsQuery();
   const { queueToast } = useQueuedToast();
-  const [timeHorizon, setTimeHorizon] = useState("4");
+  const {
+    isLoading,
+    isError,
+    error,
+    settings,
+    timeHorizonLabel,
+    onTimeHorizonChange,
+    recentTransactions,
+    tierData,
+    memberGrowthChartData,
+    revenueChartData,
+    summary,
+  } = useOverviewTab();
 
   if (isLoading) {
     return <Box css={{ font: "caption", color: "secondary" }}>Loading…</Box>;
@@ -79,100 +60,14 @@ export function OverviewTab() {
   if (isError) {
     return (
       <Box css={{ font: "caption", color: "critical" }}>
-        {error.message ?? "Something went wrong"}
+        {error?.message ?? "Something went wrong"}
       </Box>
     );
   }
 
-  if (!data) {
+  if (!summary) {
     return null;
   }
-
-  const {
-    activity,
-    membersTrend,
-    revenueTrend,
-    redemptionsTrend,
-    pointsLiabilityTrend,
-  } = data;
-  const recentTransactions = activity.slice(0, 5);
-
-  const weeksToShow = parseInt(timeHorizon);
-  const slicedMembersTrend = membersTrend.slice(-weeksToShow);
-  const slicedRevenueTrend = revenueTrend.slice(-weeksToShow);
-
-  // Current and previous period values for growth indicators
-  const prevMembersTrend = membersTrend.slice(-weeksToShow * 2, -weeksToShow);
-  const prevRevenueTrend = revenueTrend.slice(-weeksToShow * 2, -weeksToShow);
-  const prevRedemptionsTrend = redemptionsTrend.slice(
-    -weeksToShow * 2,
-    -weeksToShow,
-  );
-  const currRedemptionsTrend = redemptionsTrend.slice(-weeksToShow);
-
-  const currentMembers =
-    slicedMembersTrend[slicedMembersTrend.length - 1]?.value ?? 0;
-  const prevMembers = prevMembersTrend[prevMembersTrend.length - 1]?.value ?? 0;
-
-  const currentRevenue = slicedRevenueTrend.reduce((s, d) => s + d.value, 0);
-  const prevRevenue = prevRevenueTrend.reduce((s, d) => s + d.value, 0);
-
-  const currentRedemptions = currRedemptionsTrend.reduce(
-    (s, d) => s + d.value,
-    0,
-  );
-  const prevRedemptions = prevRedemptionsTrend.reduce((s, d) => s + d.value, 0);
-
-  const currLiability = pointsLiabilityTrend.slice(-weeksToShow);
-  const prevLiability = pointsLiabilityTrend.slice(
-    -weeksToShow * 2,
-    -weeksToShow,
-  );
-  const currentLiabilityValue =
-    currLiability[currLiability.length - 1]?.value ?? 0;
-  const prevLiabilityValue =
-    prevLiability[prevLiability.length - 1]?.value ?? 0;
-
-  // For 12-month view, aggregate weekly data into monthly buckets
-  const aggregateMonthly = <T extends { week: string; value: number }>(
-    data: T[],
-    mode: "sum" | "last",
-  ) => {
-    const withYears = assignYears(data);
-    const months = new Map<string, number[]>();
-    for (const d of withYears) {
-      const key = d.date.toLocaleDateString("en-US", {
-        month: "short",
-        year: "2-digit",
-      });
-      const existing = months.get(key);
-      if (existing) {
-        existing.push(d.value);
-      } else {
-        months.set(key, [d.value]);
-      }
-    }
-    return Array.from(months.entries()).map(([month, values]) => ({
-      label: month,
-      value:
-        mode === "sum"
-          ? values.reduce((a, b) => a + b, 0)
-          : values[values.length - 1],
-    }));
-  };
-
-  const isYearView = weeksToShow === 52;
-  const filteredMembersTrend = isYearView
-    ? aggregateMonthly(slicedMembersTrend, "last")
-    : slicedMembersTrend;
-  const filteredRevenueTrend = isYearView
-    ? aggregateMonthly(slicedRevenueTrend, "sum")
-    : slicedRevenueTrend;
-
-  const tierData = MEMBER_TIERS.map((tier) => ({
-    label: tier,
-    value: members?.filter((member) => member.tier === tier).length ?? 0,
-  }));
 
   return (
     <OverviewPage
@@ -180,7 +75,7 @@ export function OverviewTab() {
         <>
           <PageModule title="Program summary">
             <Menu
-              onAction={(key) => setTimeHorizon(String(key))}
+              onAction={(key) => onTimeHorizonChange(String(key))}
               trigger={
                 <Link>
                   <Box
@@ -204,11 +99,7 @@ export function OverviewTab() {
                     </Inline>
                     <Inline css={{ color: "secondary" }}>|</Inline>
                     <Inline css={{ fontWeight: "semibold" }}>
-                      {timeHorizon === "4"
-                        ? "Last 4 weeks"
-                        : timeHorizon === "12"
-                          ? "Last 12 weeks"
-                          : "Last 12 months"}
+                      {timeHorizonLabel}
                     </Inline>
                   </Box>
                 </Link>
@@ -238,11 +129,9 @@ export function OverviewTab() {
                 </Inline>
                 <Box css={{ stack: "x", gap: "xsmall", alignY: "baseline" }}>
                   <Inline css={{ font: "subtitle", fontWeight: "bold" }}>
-                    {currentMembers}
+                    {summary.totalMembers.value}
                   </Inline>
-                  <GrowthIndicator
-                    pct={pctChange(currentMembers, prevMembers)}
-                  />
+                  <GrowthIndicator pct={summary.totalMembers.pct} />
                 </Box>
               </Box>
               <Box css={{ stack: "y", gap: "xxsmall" }}>
@@ -251,11 +140,9 @@ export function OverviewTab() {
                 </Inline>
                 <Box css={{ stack: "x", gap: "xsmall", alignY: "baseline" }}>
                   <Inline css={{ font: "subtitle", fontWeight: "bold" }}>
-                    {formatCurrency(currentRevenue, settings?.currency)}
+                    {formatCurrency(summary.revenue.value, settings?.currency)}
                   </Inline>
-                  <GrowthIndicator
-                    pct={pctChange(currentRevenue, prevRevenue)}
-                  />
+                  <GrowthIndicator pct={summary.revenue.pct} />
                 </Box>
               </Box>
               <Box css={{ stack: "y", gap: "xxsmall" }}>
@@ -264,11 +151,9 @@ export function OverviewTab() {
                 </Inline>
                 <Box css={{ stack: "x", gap: "xsmall", alignY: "baseline" }}>
                   <Inline css={{ font: "subtitle", fontWeight: "bold" }}>
-                    {currentRedemptions}
+                    {summary.redemptions.value}
                   </Inline>
-                  <GrowthIndicator
-                    pct={pctChange(currentRedemptions, prevRedemptions)}
-                  />
+                  <GrowthIndicator pct={summary.redemptions.pct} />
                 </Box>
               </Box>
               <Box css={{ stack: "y", gap: "xxsmall" }}>
@@ -278,22 +163,16 @@ export function OverviewTab() {
                 <Box css={{ stack: "x", gap: "xsmall", alignY: "baseline" }}>
                   <Inline css={{ font: "subtitle", fontWeight: "bold" }}>
                     {formatCurrency(
-                      pointsToDollars(
-                        currentLiabilityValue,
-                        settings?.pointsPerDollar,
-                      ),
+                      summary.pointsLiability.value,
                       settings?.currency,
                     )}
                   </Inline>
-                  <GrowthIndicator
-                    pct={pctChange(currentLiabilityValue, prevLiabilityValue)}
-                  />
+                  <GrowthIndicator pct={summary.pointsLiability.pct} />
                 </Box>
               </Box>
             </Box>
           </PageModule>
 
-          {/* Charts in container card */}
           <Box
             css={{
               stack: "y",
@@ -316,26 +195,7 @@ export function OverviewTab() {
                 Member growth
               </Inline>
               <Box css={{ height: 182 }}>
-                <LineChart
-                  data={
-                    isYearView
-                      ? (
-                          filteredMembersTrend as {
-                            label: string;
-                            value: number;
-                          }[]
-                        ).map((d) => ({
-                          x: new Date(`1 ${d.label}`),
-                          y: d.value,
-                          name: "Members",
-                        }))
-                      : assignYears(slicedMembersTrend).map((d) => ({
-                          x: d.date,
-                          y: d.value,
-                          name: "Members",
-                        }))
-                  }
-                />
+                <LineChart data={memberGrowthChartData} />
               </Box>
             </Box>
 
@@ -352,26 +212,7 @@ export function OverviewTab() {
                 Revenue from members
               </Inline>
               <Box css={{ height: 182 }}>
-                <BarChart
-                  data={
-                    isYearView
-                      ? (
-                          filteredRevenueTrend as {
-                            label: string;
-                            value: number;
-                          }[]
-                        ).map((d) => ({
-                          x: d.label,
-                          y: d.value,
-                          name: "Revenue",
-                        }))
-                      : slicedRevenueTrend.map((d) => ({
-                          x: d.week,
-                          y: d.value,
-                          name: "Revenue",
-                        }))
-                  }
-                />
+                <BarChart data={revenueChartData} />
               </Box>
             </Box>
           </Box>
