+++ examples/full-page/src/pages/Home/tabs/RewardsTab/index.tsx
@@ -1,0 +1,134 @@
+import { Box, Button, Inline, Link } from "@stripe/ui-extension-sdk/ui";
+import { DataTable } from "@stripe/ui-extension-sdk/ui/experimental";
+import { useRoute } from "@stripe/ui-extension-sdk/navigation";
+import { FilterSelect } from "@/components/FilterSelect";
+import { useRewardsTab } from "./useRewardsTab";
+
+interface RewardsTabProps {
+  onEdit: (rewardId: string) => void;
+}
+
+export function RewardsTab({ onEdit }: RewardsTabProps) {
+  const { setRoute } = useRoute();
+  const {
+    items,
+    filterCategory,
+    filterStatus,
+    hasActiveFilters,
+    tableKey,
+    selectedCount,
+    hasSelection,
+    isArchivingBatch,
+    isLoading,
+    isError,
+    error,
+    onFilterCategory,
+    onFilterStatus,
+    onClearFilters,
+    onBatchChange,
+    onBatchArchive,
+    onArchiveReward,
+  } = useRewardsTab();
+
+  if (isLoading) {
+    return <Box css={{ font: "caption", color: "secondary" }}>Loading…</Box>;
+  }
+
+  if (isError) {
+    return (
+      <Box css={{ font: "caption", color: "critical" }}>
+        {error?.message ?? "Something went wrong"}
+      </Box>
+    );
+  }
+
+  return (
+    <Box css={{ stack: "y", gap: "medium" }}>
+      <Box
+        css={{
+          stack: "x",
+          gap: "small",
+          alignY: "center",
+          distribute: "space-between",
+        }}
+      >
+        <Box css={{ stack: "x", gap: "small", alignY: "center" }}>
+          <FilterSelect
+            label="Category"
+            value={filterCategory}
+            options={[
+              { label: "Coffee", value: "Coffee" },
+              { label: "Merchandise", value: "Merchandise" },
+            ]}
+            onChange={onFilterCategory}
+          />
+          <FilterSelect
+            label="Status"
+            value={filterStatus}
+            options={[
+              { label: "Available", value: "Available" },
+              { label: "Out of stock", value: "Out of stock" },
+            ]}
+            onChange={onFilterStatus}
+          />
+          {hasActiveFilters && (
+            <Link onPress={onClearFilters}>
+              <Inline css={{ fontWeight: "semibold" }}>Clear filters</Inline>
+            </Link>
+          )}
+        </Box>
+        {hasSelection && (
+          <Button
+            type="destructive"
+            onPress={onBatchArchive}
+            pending={isArchivingBatch}
+          >
+            Archive ({selectedCount})
+          </Button>
+        )}
+      </Box>
+      <DataTable
+        key={tableKey}
+        pagination={{
+          pageSize: 15,
+        }}
+        batchable={{
+          onBatchChange,
+        }}
+        columns={[
+          { key: "name", label: "Reward" },
+          { key: "description", label: "Description" },
+          { key: "category", label: "Category" },
+          { key: "pointsCost", label: "Points cost" },
+          { key: "redemptionCount", label: "Redemptions" },
+          {
+            key: "status",
+            label: "Status",
+            cell: {
+              type: "status",
+              statusMap: {
+                Available: "positive",
+                "Out of stock": "neutral",
+              },
+            },
+          },
+        ]}
+        items={items}
+        onRowClick={(item) => setRoute("reward", { id: String(item.id) })}
+        rowActions={[
+          {
+            id: "edit",
+            label: "Edit",
+            onPress: (item) => onEdit(String(item.id)),
+          },
+          {
+            id: "archive",
+            label: "Archive",
+            type: "destructive",
+            onPress: (item) => onArchiveReward(String(item.id)),
+          },
+        ]}
+      />
+    </Box>
+  );
+}
